Move upload forms from users to core app
