+++ src/core/jinja2/core/upload-artwork.jinja2
