--- src/core/jinja2/core/upload.jinja2
@@ -1,11 +1,0 @@
-{% extends '/core/base.jinja2' %}
-
-{% block content %}
-
-    <form action="{{url('upload-image')}}" method="post" enctype="multipart/form-data">
-        {{ csrf_input }}
-        {{ form }}
-        <input type="submit" value="Upload" />
-    </form>
-
-{% endblock %}
